+++ CircleRoom/Stage/Stage1_3.cpp
@@ -22,7 +22,7 @@ namespace
 	// ’Êí•¶š—ñ‚ÌF
 	constexpr unsigned int kWhiteColor = 0xf0ece5;
 	// ¶¬ŠÔŠuƒtƒŒ[ƒ€
-	constexpr int kCreateFrame = 60 * 5.5;
+	constexpr int kCreateFrame = static_cast<int>(60 * 5.5);
 
 	// ¶¬”
 	constexpr int kCreateNum = 2;
